+++ Modules/xStorage/Examples/Resources/xDisk/2-xDisk_InitializeDataDiskUsingUniqueId.ps1
@@ -14,7 +14,7 @@ Configuration Example
     {
         xWaitforDisk Disk2
         {
-             DiskId = '5E1E50A401000000001517FFFF0AEB84' # Disk 2
+             DiskNumber = '5E1E50A401000000001517FFFF0AEB84' # Disk 2
              DiskIdType = 'UniqueId'
              RetryIntervalSec = 60
              RetryCount = 60
@@ -22,7 +22,7 @@ Configuration Example
 
         xDisk GVolume
         {
-             DiskId = '5E1E50A401000000001517FFFF0AEB84' # Disk 2
+             DiskNumber = '5E1E50A401000000001517FFFF0AEB84' # Disk 2
              DiskIdType = 'UniqueId'
              DriveLetter = 'G'
              Size = 10GB
@@ -31,7 +31,7 @@ Configuration Example
 
         xDisk JVolume
         {
-             DiskId = '5E1E50A401000000001517FFFF0AEB84' # Disk 2
+             DiskNumber = '5E1E50A401000000001517FFFF0AEB84' # Disk 2
              DiskIdType = 'UniqueId'
              DriveLetter = 'J'
              FSLabel = 'Data'
@@ -40,7 +40,7 @@ Configuration Example
 
         xWaitforDisk Disk3
         {
-             DiskId = '5E1E50A4010000000029AB39450AC9A5' # Disk 3
+             DiskNumber = '5E1E50A4010000000029AB39450AC9A5' # Disk 3
              DiskIdType = 'UniqueId'
              RetryIntervalSec = 60
              RetryCount = 60
@@ -48,7 +48,7 @@ Configuration Example
 
         xDisk SVolume
         {
-             DiskId = '5E1E50A4010000000029AB39450AC9A5' # Disk 3
+             DiskNumber = '5E1E50A4010000000029AB39450AC9A5' # Disk 3
              DiskIdType = 'UniqueId'
              DriveLetter = 'S'
              Size = 100GB